requirements.txt
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ConfigArgParse==1.7.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.1
locust-cloud==1.27.12
lru-dict==1.3.0
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
prometheus_client==0.23.1
protobuf==6.33.0
psutil==7.1.2
pyee==13.0.0
Pygments==2.19.2
pysyncobj==0.3.14
pytest==8.4.2
python-engineio==4.12.3
python-socketio==5.14.3
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.0.1
requests==2.32.4
setuptools==80.9.0
simple-websocket==1.1.0
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.0
zope.interface==8.0.1
